Move torch policies
