+++ examples/torch/tsac_halfcheetah.py
@@ -7,7 +7,7 @@ import numpy as np
 import garage.torch.algos.pytorch_util as ptu
 from garage.torch.envs.wrappers import NormalizedBoxEnv
 from garage.torch.launchers.launcher_util import setup_logger
-from garage.torch.algos.sac.policies import TanhGaussianPolicy
+from garage.torch.policies import TanhGaussianPolicy
 from garage.torch.algos.networks import FlattenMlp
 from garage.torch.algos.sac.twin_sac import TwinSAC
