+++ garage/torch/algos/sac/sac.py
@@ -5,7 +5,7 @@ from torch import nn as nn
 import garage.torch.algos.pytorch_util as ptu
 from garage.torch.core.eval_util import create_stats_ordered_dict
 from garage.torch.algos.torch_rl_algorithm import TorchRLAlgorithm
-from garage.torch.algos.sac.policies import MakeDeterministic
+from garage.torch.policies import MakeDeterministic
 
 
 class SoftActorCritic(TorchRLAlgorithm):
