+++ garage/torch/algos/sac/twin_sac.py
@@ -4,7 +4,7 @@ import torch.optim as optim
 
 import garage.torch.algos.pytorch_util as ptu
 from garage.torch.core.eval_util import create_stats_ordered_dict
-from garage.torch.algos.sac.policies import MakeDeterministic
+from garage.torch.policies import MakeDeterministic
 from garage.torch.algos.torch_rl_algorithm import TorchRLAlgorithm
