+++ garage/torch/policies/__init__.py
@@ -1,0 +1,4 @@
+from garage.torch.policies.policies import TanhGaussianPolicy
+from garage.torch.policies.policies import MakeDeterministic
+
+__all__ = ["TanhGaussianPolicy", "MakeDeterministic"]
